fix(keymaxxer): remove unsupported secret-removal integration
Drop keymaxxer_rm/removeSecret, the removeRepositoryGitHubToken mutation,
and the remove-github-token CLI. Token deletion stays a Keymaxxer CLI
operation.

Closes #311
