+++ apps/harness/test/keymaxxer-github-layer.test.ts
@@ -24,7 +24,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
           addCalled = true
           return true
         }),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({ exitCode: 0, stdout: "[]", stderr: "" })
@@ -60,7 +59,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({ exitCode: 0, stdout: "[]", stderr: "" })
@@ -95,7 +93,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({
@@ -189,7 +186,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({
@@ -251,7 +247,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({ exitCode: 0, stdout: "321", stderr: "" })
@@ -363,7 +358,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({
@@ -401,7 +395,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: () =>
         Effect.succeed({
           exitCode: 1,
@@ -441,7 +434,6 @@ describe("Keymaxxer-backed GitHub layer", () => {
       findSecrets: () => Effect.die("not used"),
       hasSecret: () => Effect.die("not used"),
       addSecret: () => Effect.die("not used"),
-      removeSecret: () => Effect.die("not used"),
       runWithSecrets: (input) => {
         runs.push(input)
         return Effect.succeed({
