+++ apps/ready-for-agent/src/domain.ts
@@ -5,3 +5,12 @@ export interface LocalRepository {
   readonly isBare: boolean
   readonly paused: true
 }
+
+export type RepositorySummary = {
+  readonly id: string
+  readonly githubOwner: string
+  readonly githubRepo: string
+  readonly localPath: string
+  readonly isBare: boolean
+  readonly paused: boolean
+}
