+++ apps/ready-for-agent/src/services/graphql-api.ts
@@ -1,21 +1,13 @@
 import { Context, Effect, Layer, Schema } from "effect"
 import { createClient } from "@ready-for-agent/graphql-client"
-import type { LocalRepository } from "../domain.ts"
+import type { LocalRepository, RepositorySummary } from "../domain.ts"
 import { formatGraphqlRequestFailure } from "../graphql-error.ts"
 import { resolveGraphqlUrl } from "../graphql-url.ts"
-import type { RepositorySummary } from "../resolve-repository-target.ts"
 
 export class GraphqlRequestFailed extends Schema.TaggedErrorClass<GraphqlRequestFailed>()(
   "GraphqlRequestFailed",
   { message: Schema.String },
 ) {}
-
-export type RepositoryCredentialSummary = {
-  readonly repositoryId: string
-  readonly configured: boolean
-  readonly githubTokenSecretName: string
-  readonly githubTokenCreationUrl: string
-}
 
 export class GraphqlApi extends Context.Service<
   GraphqlApi,
@@ -23,13 +15,6 @@ export class GraphqlApi extends Context.Service<
     readonly addRepository: (
       repository: LocalRepository,
     ) => Effect.Effect<RepositorySummary, GraphqlRequestFailed>
-    readonly listRepositories: Effect.Effect<
-      readonly RepositorySummary[],
-      GraphqlRequestFailed
-    >
-    readonly removeRepositoryGitHubToken: (
-      repositoryId: string,
-    ) => Effect.Effect<RepositoryCredentialSummary, GraphqlRequestFailed>
   }
 >()("ready-for-agent/GraphqlApi") {
   static readonly layer = Layer.succeed(GraphqlApi, {
@@ -62,54 +47,6 @@ export class GraphqlApi extends Context.Service<
             throw new Error("addRepository returned null")
           }
           return added
-        },
-        catch: (cause) =>
-          new GraphqlRequestFailed({
-            message: formatGraphqlRequestFailure(cause),
-          }),
-      }),
-    listRepositories: Effect.tryPromise({
-      try: async () => {
-        const client = createClient({
-          url: resolveGraphqlUrl(),
-        })
-        const result = await client.query({
-          repositories: {
-            id: true,
-            githubOwner: true,
-            githubRepo: true,
-            localPath: true,
-            isBare: true,
-            paused: true,
-          },
-        })
-        return result.repositories ?? []
-      },
-      catch: (cause) =>
-        new GraphqlRequestFailed({
-          message: formatGraphqlRequestFailure(cause),
-        }),
-    }),
-    removeRepositoryGitHubToken: (repositoryId) =>
-      Effect.tryPromise({
-        try: async () => {
-          const client = createClient({
-            url: resolveGraphqlUrl(),
-          })
-          const result = await client.mutation({
-            removeRepositoryGitHubToken: {
-              __args: { repositoryId },
-              repositoryId: true,
-              configured: true,
-              githubTokenSecretName: true,
-              githubTokenCreationUrl: true,
-            },
-          })
-          const removed = result.removeRepositoryGitHubToken
-          if (!removed) {
-            throw new Error("removeRepositoryGitHubToken returned null")
-          }
-          return removed
         },
         catch: (cause) =>
           new GraphqlRequestFailed({
