+++ docs/adr/0023-public-npm-binary-distribution.md
@@ -1,3 +1,3 @@
 # Public npm binary distribution
 
-The Harness is distributed as a single public unscoped package `ready-for-agent` whose bin boots the full local app (UI + backend), not as a multi-package publish of every workspace library. Runtime is a **compiled Bun binary** so end users need not install Bun; a thin JS launcher plus **platform packages as `optionalDependencies`** (`ready-for-agent-linux-x64`, `ready-for-agent-darwin-arm64`, …) installs only the host binary (esbuild-style). First platforms: linux and darwin × x64/arm64; Windows later. Host tools remain external: `git`, `gh`, and OpenCode on PATH; Keymaxxer is optional via `KEYMAXXER_ENTRYPOINT` or `keymaxxer` on PATH—no hardcoded developer contrib path. Operator CLI commands (`add`, `remove-github-token`, …) live on the same binary; the separate `apps/cli` product is retired. Publishable sources live under `apps/ready-for-agent` plus platform packages. Rejected: Node-native rewrite for v1, fat multi-OS tarball, postinstall download from GitHub, dual npm packages (harness + cli), and add-repository as a web form instead of the binary.
+The Harness is distributed as a single public unscoped package `ready-for-agent` whose bin boots the full local app (UI + backend), not as a multi-package publish of every workspace library. Runtime is a **compiled Bun binary** so end users need not install Bun; a thin JS launcher plus **platform packages as `optionalDependencies`** (`ready-for-agent-linux-x64`, `ready-for-agent-darwin-arm64`, …) installs only the host binary (esbuild-style). First platforms: linux and darwin × x64/arm64; Windows later. Host tools remain external: `git`, `gh`, and OpenCode on PATH; Keymaxxer is optional via `KEYMAXXER_ENTRYPOINT` or `keymaxxer` on PATH—no hardcoded developer contrib path. Operator CLI commands (`add`, …) live on the same binary; the separate `apps/cli` product is retired. Publishable sources live under `apps/ready-for-agent` plus platform packages. Rejected: Node-native rewrite for v1, fat multi-OS tarball, postinstall download from GitHub, dual npm packages (harness + cli), and add-repository as a web form instead of the binary.
