+++ packages/keymaxxer-service/src/lib/service.ts
@@ -24,9 +24,6 @@ export interface KeymaxxerServiceShape {
   readonly addSecret: (
     input: AddSecretInput,
   ) => Effect.Effect<boolean, KeymaxxerError>
-  readonly removeSecret: (
-    name: SecretName,
-  ) => Effect.Effect<boolean, KeymaxxerError>
   readonly runWithSecrets: (
     input: RunWithSecretsInput,
   ) => Effect.Effect<RunWithSecretsResult, KeymaxxerError>
@@ -53,8 +50,6 @@ export const testKeymaxxerLayer = (
           secrets.add(input.name)
           return true
         }),
-      removeSecret: (name: SecretName) =>
-        Effect.sync(() => secrets.delete(name)),
       runWithSecrets: () =>
         Effect.succeed({ exitCode: 0, stdout: "", stderr: "" }),
     }
@@ -68,7 +63,6 @@ export const disabledKeymaxxerLayer: Layer.Layer<KeymaxxerService> =
     findSecret: () => Effect.succeed(null),
     findSecrets: (inputs) => Effect.succeed(inputs.map(() => null)),
     addSecret: () => Effect.succeed(false),
-    removeSecret: () => Effect.succeed(false),
     runWithSecrets: (input) =>
       Effect.tryPromise({
         try: () =>
