+++ packages/work-item-lifecycle/test/create-pr.spec.ts
@@ -65,7 +65,6 @@ const stubKeymaxxer = (
     findSecret: () => Effect.succeed("GITHUB_TOKEN_ACME_WIDGETS"),
     findSecrets: () => Effect.succeed([]),
     addSecret: () => Effect.succeed(true),
-    removeSecret: () => Effect.succeed(true),
     runWithSecrets: () =>
       Effect.succeed({ exitCode: 0, stdout: "", stderr: "" }),
     ...overrides,
