+++ packages/work-item-lifecycle/test/remove-worktree.spec.ts
@@ -34,7 +34,6 @@ const stubKeymaxxer = (
     findSecret: () => Effect.succeed("GITHUB_TOKEN_ACME_WIDGETS"),
     findSecrets: () => Effect.succeed([]),
     addSecret: () => Effect.succeed(true),
-    removeSecret: () => Effect.succeed(true),
     runWithSecrets: () =>
       Effect.succeed({ exitCode: 0, stdout: "[]", stderr: "" }),
     ...overrides,
